Fix objc message logging for 64-bit code
The Objective-C runtime version 2 (used for non-32-bit architectures)
added a runtime lock. This lock is always taken before the message log
lock, but the lockdebug subsystem wasn't made aware of this; because of
this, it would throw an error saying the lock order wasn't defined.
Let's define it so it can calm down.